Some data types are updated to unsigned long long because of address space size may be too large.

+++ functions/prime_functions.c
@@ -1,6 +1,6 @@
 #include "../headers/prime_functions.h"
 
-int is_prime_number(unsigned int number) {
+unsigned int is_prime_number(unsigned int number) {
     unsigned int divisor = 2, counter = 0;
     while (number != 1) {
         if (number % divisor == 0) {
@@ -14,7 +14,7 @@ int is_prime_number(unsigned int number) {
     return 1;
 }
 
-unsigned int first_prime_number_after(unsigned int number) {
+unsigned long long first_prime_number_after(unsigned long long number) {
     number++;
     while (!is_prime_number(number)) number++;
     return number;
